# tests/fixtures/openapi_snapshot.json
{
  "components": {
    "schemas": {
      "ApprovalDecision": {
        "description": "Решение технолога.",
        "enum": [
          "approved",
          "rejected"
        ],
        "title": "ApprovalDecision",
        "type": "string"
      },
      "ApprovalRequest": {
        "description": "Запрос на approve/reject.\n\nexpected_revision — optimistic concurrency.\nconfirmed — явное подтверждение технолога (checkbox/фразa в UI).",
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "confirmed": {
            "description": "Технолог подтверждает ручную проверку и допуск к производству",
            "title": "Confirmed",
            "type": "boolean"
          },
          "expected_revision": {
            "description": "Ожидаемая ревизия для OCC",
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision",
          "confirmed"
        ],
        "title": "ApprovalRequest",
        "type": "object"
      },
      "ApprovalResponse": {
        "description": "Успешный ответ approve.",
        "properties": {
          "audit": {
            "additionalProperties": true,
            "title": "Audit",
            "type": "object"
          },
          "decision": {
            "$ref": "#/components/schemas/ApprovalDecision"
          },
          "new_status": {
            "title": "New Status",
            "type": "string"
          },
          "order_manufacturing_status": {
            "title": "Order Manufacturing Status",
            "type": "string"
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          },
          "revision_number": {
            "title": "Revision Number",
            "type": "integer"
          }
        },
        "required": [
          "revision_id",
          "decision",
          "new_status",
          "revision_number",
          "order_manufacturing_status",
          "audit"
        ],
        "title": "ApprovalResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidateResponse": {
        "description": "Результат валидации для UI технолога.",
        "properties": {
          "blocking_errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Blocking Errors",
            "type": "array"
          },
          "order_id": {
            "format": "uuid",
            "title": "Order Id",
            "type": "string"
          },
          "revision_number": {
            "title": "Revision Number",
            "type": "integer"
          },
          "spec_valid": {
            "title": "Spec Valid",
            "type": "boolean"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "order_id",
          "revision_number",
          "spec_valid",
          "blocking_errors",
          "warnings",
          "summary"
        ],
        "title": "ValidateResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationIssue": {
        "description": "Отдельная ошибка/предупреждение с точной ссылкой на операцию.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "panel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Panel Id"
          },
          "severity": {
            "default": "error",
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ValidationIssue",
        "type": "object"
      },
...
      "_LegacyApprovalRequest": {
        "description": "Тело совместимого revision-level approve/reject запроса.",
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "decision": {
            "$ref": "#/components/schemas/ApprovalDecision"
          },
          "expected_revision": {
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "decision",
          "expected_revision"
        ],
        "title": "_LegacyApprovalRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Furniture AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/manufacturing/revisions/{revision_id}/approve": {
      "post": {
        "description": "Совместимый revision-level approve маршрут для старых клиентов.",
        "operationId": "legacy_approve_revision_api_v1_manufacturing_revisions__revision_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_LegacyApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Legacy Approve Revision Api V1 Manufacturing Revisions  Revision Id  Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Legacy Approve Revision",
        "tags": [
          "Manufacturing Approvals (legacy)"
        ]
      }
    },
    "/api/v1/manufacturing/revisions/{revision_id}/cam-gate": {
      "get": {
        "description": "Совместимый read-only gate: CAM допускается только approved revision.",
        "operationId": "legacy_cam_gate_api_v1_manufacturing_revisions__revision_id__cam_gate_get",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Legacy Cam Gate Api V1 Manufacturing Revisions  Revision Id  Cam Gate Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Legacy Cam Gate",
        "tags": [
          "Manufacturing Approvals (legacy)"
        ]
      }
    },
    "/api/v1/manufacturing/revisions/{revision_id}/reject": {
      "post": {
        "description": "Совместимый revision-level reject маршрут для старых клиентов.",
        "operationId": "legacy_reject_revision_api_v1_manufacturing_revisions__revision_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_LegacyApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Legacy Reject Revision Api V1 Manufacturing Revisions  Revision Id  Reject Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Legacy Reject Revision",
        "tags": [
          "Manufacturing Approvals (legacy)"
        ]
      }
    },
...
    "/api/v1/orders/{order_id}/manufacturing/approve": {
      "post": {
        "description": "Настоящее утверждение технологом.\n\n- expected_revision (stale → 409)\n- confirmed=true обязательно (технолог подтверждает)\n- tenant isolation + auth\n- обновляет revision.status + order.approved_manufacturing_revision\n- пишет AuditLog без каких-либо AI prompt'ов",
        "operationId": "approve_manufacturing_api_v1_orders__order_id__manufacturing_approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve Manufacturing",
        "tags": [
          "Manufacturing Approvals (Task 8)"
        ]
      }
    },
    "/api/v1/orders/{order_id}/manufacturing/validate": {
      "post": {
        "description": "Валидация текущей ревизии перед утверждением.\n\nВозвращает blocking_errors, warnings и точные ссылки на операции.\nТребует аутентификации + tenant.",
        "operationId": "validate_manufacturing_api_v1_orders__order_id__manufacturing_validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Validate Manufacturing",
        "tags": [
          "Manufacturing Approvals (Task 8)"
        ]
      }
    },
...
  }
}